refactor: decompose useDisplayState god hook
